version 1.2
Added VocaDB search

Database functionality implemented (needs further testing)
The database (database.js) currently references [all videos in this vocarank playlist](https://www.nicovideo.jp/mylist/3678045). These were obtained from grabbing the linked videos in the descriptions of the vocaranks, so there MAY be songs missing, but I was unable to find any songs that were present in the video but not the description during my review (which WAS brief)
I hope this feature will grow to be game-changing, and soon I plan to expand the database to include all vocaranks I can find that have mylists or videos linked in their descriptions.